+++ codegen/example/src-gen/main/java/org/fuin/ddd4j/codegen/example/TheEntityId.java
@@ -21,6 +21,7 @@ import org.fuin.ddd4j.core.EntityType;
 import org.fuin.ddd4j.core.HasEntityTypeConstant;
 import org.fuin.ddd4j.core.StringBasedEntityType;
 import org.fuin.objects4j.common.ConstraintViolationException;
+import org.fuin.objects4j.common.HasPublicStaticIsValidMethod;
 import org.fuin.objects4j.common.HasPublicStaticValueOfMethod;
 
 import javax.annotation.concurrent.Immutable;
@@ -35,6 +36,8 @@ import javax.annotation.concurrent.Immutable;
 @Immutable
 @HasPublicStaticValueOfMethod
 @HasEntityTypeConstant
+@HasPublicStaticIsValidMethod
+@HasPublicStaticIsValidMethod(param = Integer.class)
 
 public final class TheEntityId extends IntegerEntityId {